Update JSON Schema Converter to support int64 values (#315)
This pull request allows the json schema converter to support 64 bit
values (previously only supported up to 32 bit), which would cause
serving engines to crash in production when values exceeded 32 bits.

+++ cpp/json_schema_converter.h
@@ -70,7 +70,7 @@ std::string JSONSchemaToEBNF(
  * \param end The end of the range (inclusive). If null assume infinity.
  * \returns The regex pattern that matches integers/floats in the given range.
  */
-std::string GenerateRangeRegex(std::optional<int> start, std::optional<int> end);
+std::string GenerateRangeRegex(std::optional<int64_t> start, std::optional<int64_t> end);
 
 std::string GenerateFloatRangeRegex(std::optional<double> start, std::optional<double> end);
